Refactors Directory and Segmenter
Moves `global_depth` to be the responsibility of the Directory instead
of tracking it in the Segmenter. The Segmenter instead now has a generic
Header associated type that can by set to serializable type. The
BasicFileSegmenter simply has a u64 Header type as the only extra data
it persists (in addition to the Segments themselves) is `num_segments`.
Also refactors the segment allocate methods to return the index of the
newly allocated segment, as it's a bit difficult to include that in the
`Segment` struct directly (due to it not being a serialized property and
using the Serialized trait). Also refactors the index to be a u64 for
consistency. This can maybe be a significant space savings in the
Directory to only support u32, but for now it simplifies the code.