Refactors encoded data types
Changes all depth (local and global) to be u8, as their max possible
value is current 64. Effectively establishes a max number of segments by
changing the directory entry size to u32. This halves the size of the
directory in-memory/on-disk and still allows for a sizeable number of
segments, almost certainly beyond what is actually practical with
today's hardware.

Also begins implementing a memory mapped directory.